+++ webpack.mix.js
@@ -25,7 +25,7 @@ if (!mix.inProduction()) {
 
 mix
   .copy('./node_modules/font-awesome/fonts/', 'public/fonts')
-  .copy('./node_modules/simple-line-icons/fonts/', 'public/fonts')
+  .copy('./node_modules/orchid-icons/src/fonts/', 'public/fonts')
   .copyDirectory('./node_modules/tinymce', 'public/js/tinymce')
   .sass('resources/assets/sass/app.scss', 'css/orchid.css')
   .js('resources/assets/js/app.js', 'js/orchid.js');