+++ resources/assets/js/components/menu.js
@@ -88,11 +88,16 @@ document.addEventListener('turbolinks:load', function() {
         $('#menuEdit').modal('hide');
         menu.send();
       },
+      destroy: function (id) {
+          axios
+              .delete(dashboard.prefix('/systems/menu/' + id))
+              .then(function(response) {});
+      },
       remove: function() {
         $('li[data-id=' + this.id + ']').remove();
         $('#menuEdit').modal('hide');
+        this.destroy(this.id);
         this.clear();
-        this.send();
       },
       clear: function() {
         this.label = '';