+++ src/Platform/Kernel/Dashboard.php
@@ -13,7 +13,7 @@ class Dashboard
     /**
      * ORCHID Version.
      */
-    const VERSION = '2.1';
+    const VERSION = '2.1.0';
 
     /**
      * @var