+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+
+## [2.1.0] - 2018-02-09
+
 ### Added
 - "DIV" Layouts
 - "Select" field
@@ -21,6 +24,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Demo "behaviors" are no longer published
 - Access validation does not create multiple database queries
 - Hide forms switching with their small number
+- Changing the menu, no longer changing the recording number
+- Color pallet for graphs
+- Design of the file manager
 
 ### Deprecated
 - Record string/array for fields. Use `Field::make`